Added default value of modal in complain list
